feat: add hint icon to Additional Notes field
This commit adds a hint icon to the Additional Notes field in the Job Openings resource. The hint icon provides a tooltip that explains that the field will display in the career job portal. This will help users understand the purpose of the field and how it will be used.